fix(): cleanup ai search api login by reusing same function for search and chat
