+++ src/pages/api/ai-chat.ts
@@ -1,41 +1,8 @@
 import type { APIRoute } from 'astro';
-import {
-  CLOUDFLARE_ACCOUNT_ID,
-  CLOUDFLARE_AI_API_TOKEN,
-  CLOUDFLARE_AI_SEARCH_INSTANCE_NAME,
-} from 'astro:env/server';
+import { chatCompletion, isAiSearchConfigured, jsonError, type ChatMessage } from '~/lib/ai-search-api';
 import { HISTORY_CAP } from '~/lib/ai-stream';
 
 export const prerender = false;
-
-type ChatRole = 'user' | 'assistant';
-type ChatMessage = { role: ChatRole; content: string };
-
-const jsonError = (message: string, status: number) =>
-  new Response(JSON.stringify({ error: message }), {
-    status,
-    headers: { 'Content-Type': 'application/json', 'Cache-Control': 'no-store' },
-  });
-
-type ChunkMeta = { url?: string; title?: string };
-type ChatCompletionResponse = {
-  choices?: { message?: { content?: string } }[];
-  // Non-streaming response keeps metadata under `item.metadata`; streaming SSE uses `item.attributes`.
-  chunks?: { item?: { metadata?: ChunkMeta; attributes?: ChunkMeta } }[];
-};
-
-const reshapeAsJson = (payload: ChatCompletionResponse) => {
-  const answer = payload.choices?.[0]?.message?.content ?? '';
-  const seen = new Set<string>();
-  const sources = (payload.chunks ?? []).flatMap((chunk) => {
-    const meta = chunk.item?.metadata ?? chunk.item?.attributes;
-    const url = meta?.url;
-    if (!url || seen.has(url)) return [];
-    seen.add(url);
-    return [{ url, title: meta?.title ?? url }];
-  });
-  return { answer, sources };
-};
 
 const isValidMessage = (message: unknown): message is ChatMessage => {
   if (typeof message !== 'object' || message === null) return false;
@@ -56,7 +23,7 @@ const capHistory = (messages: ChatMessage[]) => {
 };
 
 export const POST: APIRoute = async ({ request, url }) => {
-  if (!CLOUDFLARE_ACCOUNT_ID || !CLOUDFLARE_AI_API_TOKEN || !CLOUDFLARE_AI_SEARCH_INSTANCE_NAME) {
+  if (!isAiSearchConfigured()) {
     return jsonError('AI Search is not configured on this deployment.', 503);
   }
 
@@ -81,34 +48,5 @@ export const POST: APIRoute = async ({ request, url }) => {
 
   const wantsJson = url.searchParams.get('format') === 'json';
 
-  const upstream = await fetch(
-    `https://api.cloudflare.com/client/v4/accounts/${CLOUDFLARE_ACCOUNT_ID}/ai-search/instances/${CLOUDFLARE_AI_SEARCH_INSTANCE_NAME}/chat/completions`,
-    {
-      method: 'POST',
-      headers: {
-        Authorization: `Bearer ${CLOUDFLARE_AI_API_TOKEN}`,
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({ messages, stream: !wantsJson }),
-    },
-  );
-
-  if (wantsJson) {
-    if (!upstream.ok) {
-      return jsonError(`Upstream error (${upstream.status})`, upstream.status);
-    }
-    const payload = (await upstream.json()) as ChatCompletionResponse;
-    return new Response(JSON.stringify(reshapeAsJson(payload), null, 2), {
-      status: 200,
-      headers: { 'Content-Type': 'application/json', 'Cache-Control': 'no-store' },
-    });
-  }
-
-  return new Response(upstream.body, {
-    status: upstream.status,
-    headers: {
-      'Content-Type': upstream.headers.get('Content-Type') ?? 'text/event-stream',
-      'Cache-Control': 'no-store',
-    },
-  });
+  return chatCompletion(messages, wantsJson);
 };
